Conexión base de datos

+++ src/main/java/com/hortalsoft/products/domain/repository/ProductRepository.java
@@ -3,8 +3,7 @@ package com.hortalsoft.products.domain.repository;
 import com.hortalsoft.products.domain.entity.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<ProductEntity,Long> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
     boolean existsByName(String name);
-    boolean existsById(long id);
-
+    boolean existsById(int id);
 }
