+++ src/main/java/com/hortalsoft/products/domain/repository/SubcategoryRepository.java
@@ -3,5 +3,5 @@ package com.hortalsoft.products.domain.repository;
 import com.hortalsoft.products.domain.entity.SubcategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubcategoryRepository extends JpaRepository<SubcategoryEntity,Long> {
+public interface SubcategoryRepository extends JpaRepository<SubcategoryEntity,Integer> {
 }
